+++ include/RangeInput.h
@@ -5,11 +5,11 @@
 #include "Misc/Callbacks.h"
 #include "InputManager.h"
 
-/// Class representing an input able to make a callback to a function with a double as a parameter when triggered.
+/// Class representing an input able to make a callback to a function with a float as a parameter when triggered.
 class RangeInput
 {
     public:
-    RangeInput(const InputManager& inputManager, Callback<double>* callback);
+    RangeInput(const InputManager& inputManager, Callback<float>* callback);
     RangeInput(const RangeInput&) = delete;
     RangeInput(RangeInput&&) = delete;
     virtual ~RangeInput();
@@ -21,14 +21,14 @@ class RangeInput
 
 protected:
     const InputManager& m_inputManager;
-    Callback<double>* m_callback;
+    Callback<float>* m_callback;
 };
 
-/// Class representing a joystick axis able to make a callback to a function with a double as a parameter when moved.
+/// Class representing a joystick axis able to make a callback to a function with a float as a parameter when moved.
 class JoystickAxisBidirectionalRangeInput final : public RangeInput
 {
 public:
-    JoystickAxisBidirectionalRangeInput(const InputManager& inputManager, Callback<double>* callback,
+    JoystickAxisBidirectionalRangeInput(const InputManager& inputManager, Callback<float>* callback,
                                         unsigned int joystick, sf::Joystick::Axis axis);
     JoystickAxisBidirectionalRangeInput(const JoystickAxisBidirectionalRangeInput&) = delete;
     JoystickAxisBidirectionalRangeInput(JoystickAxisBidirectionalRangeInput&&) = delete;
@@ -43,11 +43,11 @@ private:
     mutable float m_lastAxisPosition;
 };
 
-/// Class representing a joystick axis able to make a callback to a function with a double as a parameter when moved.
+/// Class representing a joystick axis able to make a callback to a function with a float as a parameter when moved.
 class JoystickAxisUnidirectionalRangeInput final : public RangeInput
 {
 public:
-    JoystickAxisUnidirectionalRangeInput(const InputManager& inputManager, Callback<double>* callback, unsigned int joystick,
+    JoystickAxisUnidirectionalRangeInput(const InputManager& inputManager, Callback<float>* callback, unsigned int joystick,
                                          sf::Joystick::Axis axis, bool isRestrictedToPositives);
     JoystickAxisUnidirectionalRangeInput(const JoystickAxisUnidirectionalRangeInput&) = delete;
     JoystickAxisUnidirectionalRangeInput(JoystickAxisUnidirectionalRangeInput&&) = delete;
@@ -63,11 +63,11 @@ private:
     mutable float m_lastAxisPosition;
 };
 
-/// Class representing a vertical mouse wheel able to make a callback to a function with a double as a parameter when scrolled.
+/// Class representing a vertical mouse wheel able to make a callback to a function with a float as a parameter when scrolled.
 class VerticalMouseWheelBidirectionalRangeInput final : public RangeInput
 {
 public:
-    VerticalMouseWheelBidirectionalRangeInput(const InputManager& inputManager, Callback<double>* callback);
+    VerticalMouseWheelBidirectionalRangeInput(const InputManager& inputManager, Callback<float>* callback);
     VerticalMouseWheelBidirectionalRangeInput(const VerticalMouseWheelBidirectionalRangeInput&) = delete;
     VerticalMouseWheelBidirectionalRangeInput(VerticalMouseWheelBidirectionalRangeInput&&) = delete;
     VerticalMouseWheelBidirectionalRangeInput& operator=(const VerticalMouseWheelBidirectionalRangeInput&) = delete;
@@ -79,11 +79,11 @@ private:
     mutable float m_lastVertScroll;
 };
 
-/// Class representing a vertical mouse wheel able to make a callback to a function with a double as a parameter when scrolled.
+/// Class representing a vertical mouse wheel able to make a callback to a function with a float as a parameter when scrolled.
 class VerticalMouseWheelUnidirectionalRangeInput final : public RangeInput
 {
 public:
-    VerticalMouseWheelUnidirectionalRangeInput(const InputManager& inputManager, Callback<double>* callback, bool isRestrictedToPositives);
+    VerticalMouseWheelUnidirectionalRangeInput(const InputManager& inputManager, Callback<float>* callback, bool isRestrictedToPositives);
     VerticalMouseWheelUnidirectionalRangeInput(const VerticalMouseWheelUnidirectionalRangeInput&) = delete;
     VerticalMouseWheelUnidirectionalRangeInput(VerticalMouseWheelUnidirectionalRangeInput&&) = delete;
     VerticalMouseWheelUnidirectionalRangeInput& operator=(const VerticalMouseWheelUnidirectionalRangeInput&) = delete;
@@ -96,11 +96,11 @@ private:
     mutable float m_lastVertScroll;
 };
 
-/// Class representing a horizontal mouse wheel able to make a callback to a function with a double as a parameter when scrolled.
+/// Class representing a horizontal mouse wheel able to make a callback to a function with a float as a parameter when scrolled.
 class HorizontalMouseWheelBidirectionalRangeInput final : public RangeInput
 {
 public:
-    HorizontalMouseWheelBidirectionalRangeInput(const InputManager& inputManager, Callback<double>* callback);
+    HorizontalMouseWheelBidirectionalRangeInput(const InputManager& inputManager, Callback<float>* callback);
     HorizontalMouseWheelBidirectionalRangeInput(const HorizontalMouseWheelBidirectionalRangeInput&) = delete;
     HorizontalMouseWheelBidirectionalRangeInput(HorizontalMouseWheelBidirectionalRangeInput&&) = delete;
     HorizontalMouseWheelBidirectionalRangeInput& operator=(const HorizontalMouseWheelBidirectionalRangeInput&) = delete;
@@ -112,11 +112,11 @@ private:
     mutable float m_lastHorizScroll;
 };
 
-/// Class representing a horizontal mouse wheel able to make a callback to a function with a double as a parameter when scrolled.
+/// Class representing a horizontal mouse wheel able to make a callback to a function with a float as a parameter when scrolled.
 class HorizontalMouseWheelUnidirectionalRangeInput final : public RangeInput
 {
 public:
-    HorizontalMouseWheelUnidirectionalRangeInput(const InputManager& inputManager, Callback<double>* callback, bool isRestrictedToPositives);
+    HorizontalMouseWheelUnidirectionalRangeInput(const InputManager& inputManager, Callback<float>* callback, bool isRestrictedToPositives);
     HorizontalMouseWheelUnidirectionalRangeInput(const HorizontalMouseWheelUnidirectionalRangeInput&) = delete;
     HorizontalMouseWheelUnidirectionalRangeInput(HorizontalMouseWheelUnidirectionalRangeInput&&) = delete;
     HorizontalMouseWheelUnidirectionalRangeInput& operator=(const HorizontalMouseWheelUnidirectionalRangeInput&) = delete;
@@ -129,11 +129,11 @@ private:
     mutable float m_lastHorizScroll;
 };
 
-/// Class representing horizontal mouse movements able to make a callback to a function with a double as a parameter when the mouse is moved.
+/// Class representing horizontal mouse movements able to make a callback to a function with a float as a parameter when the mouse is moved.
 class VerticalMouseMovementBidirectionalRangeInput final : public RangeInput
 {
 public:
-    VerticalMouseMovementBidirectionalRangeInput(const InputManager& inputManager, Callback<double>* callback);
+    VerticalMouseMovementBidirectionalRangeInput(const InputManager& inputManager, Callback<float>* callback);
     VerticalMouseMovementBidirectionalRangeInput(const VerticalMouseMovementBidirectionalRangeInput&) = delete;
     VerticalMouseMovementBidirectionalRangeInput(VerticalMouseMovementBidirectionalRangeInput&&) = delete;
     VerticalMouseMovementBidirectionalRangeInput& operator=(const VerticalMouseMovementBidirectionalRangeInput&) = delete;
@@ -145,11 +145,11 @@ private:
     mutable float m_lastVertMouseMovement;
 };
 
-/// Class representing horizontal mouse movements able to make a callback to a function with a double as a parameter when the mouse is moved.
+/// Class representing horizontal mouse movements able to make a callback to a function with a float as a parameter when the mouse is moved.
 class VerticalMouseMovementUnidirectionalRangeInput final : public RangeInput
 {
 public:
-    VerticalMouseMovementUnidirectionalRangeInput(const InputManager& inputManager, Callback<double>* callback, bool isRestrictedToPositives);
+    VerticalMouseMovementUnidirectionalRangeInput(const InputManager& inputManager, Callback<float>* callback, bool isRestrictedToPositives);
     VerticalMouseMovementUnidirectionalRangeInput(const VerticalMouseMovementUnidirectionalRangeInput&) = delete;
     VerticalMouseMovementUnidirectionalRangeInput(VerticalMouseMovementUnidirectionalRangeInput&&) = delete;
     VerticalMouseMovementUnidirectionalRangeInput& operator=(const VerticalMouseMovementUnidirectionalRangeInput&) = delete;
@@ -162,11 +162,11 @@ private:
     mutable float m_lastVertMouseMovement;
 };
 
-/// Class representing vertical mouse movements able to make a callback to a function with a double as a parameter when the mouse is moved.
+/// Class representing vertical mouse movements able to make a callback to a function with a float as a parameter when the mouse is moved.
 class HorizontalMouseMovementBidirectionalRangeInput final : public RangeInput
 {
 public:
-    HorizontalMouseMovementBidirectionalRangeInput(const InputManager& inputManager, Callback<double>* callback);
+    HorizontalMouseMovementBidirectionalRangeInput(const InputManager& inputManager, Callback<float>* callback);
     HorizontalMouseMovementBidirectionalRangeInput(const HorizontalMouseMovementBidirectionalRangeInput&) = delete;
     HorizontalMouseMovementBidirectionalRangeInput(HorizontalMouseMovementBidirectionalRangeInput&&) = delete;
     HorizontalMouseMovementBidirectionalRangeInput& operator=(const HorizontalMouseMovementBidirectionalRangeInput&) = delete;
@@ -178,11 +178,11 @@ private:
     mutable float m_lastHorizMouseMovement;
 };
 
-/// Class representing vertical mouse movements able to make a callback to a function with a double as a parameter when the mouse is moved.
+/// Class representing vertical mouse movements able to make a callback to a function with a float as a parameter when the mouse is moved.
 class HorizontalMouseMovementUnidirectionalRangeInput final : public RangeInput
 {
 public:
-    HorizontalMouseMovementUnidirectionalRangeInput(const InputManager& inputManager, Callback<double>* callback, bool isRestrictedToPositives);
+    HorizontalMouseMovementUnidirectionalRangeInput(const InputManager& inputManager, Callback<float>* callback, bool isRestrictedToPositives);
     HorizontalMouseMovementUnidirectionalRangeInput(const HorizontalMouseMovementUnidirectionalRangeInput&) = delete;
     HorizontalMouseMovementUnidirectionalRangeInput(HorizontalMouseMovementUnidirectionalRangeInput&&) = delete;
     HorizontalMouseMovementUnidirectionalRangeInput& operator=(const HorizontalMouseMovementUnidirectionalRangeInput&) = delete;
@@ -195,11 +195,11 @@ private:
     mutable float m_lastHorizMouseMovement;
 };
 
-/// Class representing keyboard keys able to make a callback to a function with a double as a parameter when triggered.
+/// Class representing keyboard keys able to make a callback to a function with a float as a parameter when triggered.
 class KeyboardBidirectionalRangeInput final : public RangeInput
 {
 public:
-    KeyboardBidirectionalRangeInput(const InputManager& inputManager, Callback<double>* callback, sf::Keyboard::Key negativeKey, sf::Keyboard::Key positiveKey);
+    KeyboardBidirectionalRangeInput(const InputManager& inputManager, Callback<float>* callback, sf::Keyboard::Key negativeKey, sf::Keyboard::Key positiveKey);
     KeyboardBidirectionalRangeInput(const KeyboardBidirectionalRangeInput&) = delete;
     KeyboardBidirectionalRangeInput(KeyboardBidirectionalRangeInput&&) = delete;
     KeyboardBidirectionalRangeInput& operator=(const KeyboardBidirectionalRangeInput&) = delete;
@@ -212,11 +212,11 @@ private:
     sf::Keyboard::Key m_positiveKey;
 };
 
-/// Class representing keyboard keys able to make a callback to a function with a double as a parameter when triggered.
+/// Class representing keyboard keys able to make a callback to a function with a float as a parameter when triggered.
 class KeyboardUnidirectionalRangeInput final : public RangeInput
 {
 public:
-    KeyboardUnidirectionalRangeInput(const InputManager& inputManager, Callback<double>* callback, sf::Keyboard::Key key);
+    KeyboardUnidirectionalRangeInput(const InputManager& inputManager, Callback<float>* callback, sf::Keyboard::Key key);
     KeyboardUnidirectionalRangeInput(const KeyboardUnidirectionalRangeInput&) = delete;
     KeyboardUnidirectionalRangeInput(KeyboardUnidirectionalRangeInput&&) = delete;
     KeyboardUnidirectionalRangeInput& operator=(const KeyboardUnidirectionalRangeInput&) = delete;
@@ -228,11 +228,11 @@ private:
     sf::Keyboard::Key m_key;
 };
 
-/// Class representing joystick buttons able to make a callback to a function with a double as a parameter when triggered.
+/// Class representing joystick buttons able to make a callback to a function with a float as a parameter when triggered.
 class JoystickButtonBidirectionalRangeInput final : public RangeInput
 {
 public:
-    JoystickButtonBidirectionalRangeInput(const InputManager& inputManager, Callback<double>* callback, unsigned int joystick,
+    JoystickButtonBidirectionalRangeInput(const InputManager& inputManager, Callback<float>* callback, unsigned int joystick,
                                           unsigned int negativeJoystickButton, unsigned int positiveJoystickButton);
     JoystickButtonBidirectionalRangeInput(const JoystickButtonBidirectionalRangeInput&) = delete;
     JoystickButtonBidirectionalRangeInput(JoystickButtonBidirectionalRangeInput&&) = delete;
@@ -247,11 +247,11 @@ private:
     unsigned int m_positiveJoystickButton;
 };
 
-/// Class representing a joystick button able to make a callback to a function with a double as a parameter when triggered.
+/// Class representing a joystick button able to make a callback to a function with a float as a parameter when triggered.
 class JoystickButtonUnidirectionalRangeInput final : public RangeInput
 {
 public:
-    JoystickButtonUnidirectionalRangeInput(const InputManager& inputManager, Callback<double>* callback, unsigned int joystick,
+    JoystickButtonUnidirectionalRangeInput(const InputManager& inputManager, Callback<float>* callback, unsigned int joystick,
                                            unsigned int button);
     JoystickButtonUnidirectionalRangeInput(const JoystickButtonUnidirectionalRangeInput&) = delete;
     JoystickButtonUnidirectionalRangeInput(JoystickButtonUnidirectionalRangeInput&&) = delete;
@@ -265,11 +265,11 @@ private:
     unsigned int m_button;
 };
 
-/// Class representing joystick buttons able to make a callback to a function with a double as a parameter when triggered.
+/// Class representing joystick buttons able to make a callback to a function with a float as a parameter when triggered.
 class MouseButtonBidirectionalRangeInput final : public RangeInput
 {
 public:
-    MouseButtonBidirectionalRangeInput(const InputManager& inputManager, Callback<double>* callback,
+    MouseButtonBidirectionalRangeInput(const InputManager& inputManager, Callback<float>* callback,
                                        sf::Mouse::Button negativeMouseButton, sf::Mouse::Button positiveMouseButton);
     MouseButtonBidirectionalRangeInput(const MouseButtonBidirectionalRangeInput&) = delete;
     MouseButtonBidirectionalRangeInput(MouseButtonBidirectionalRangeInput&&) = delete;
@@ -283,11 +283,11 @@ private:
     sf::Mouse::Button m_positiveMouseButton;
 };
 
-/// Class representing a joystick button able to make a callback to a function with a double as a parameter when triggered.
+/// Class representing a joystick button able to make a callback to a function with a float as a parameter when triggered.
 class MouseButtonUnidirectionalRangeInput final : public RangeInput
 {
 public:
-    MouseButtonUnidirectionalRangeInput(const InputManager& inputManager, Callback<double>* callback,
+    MouseButtonUnidirectionalRangeInput(const InputManager& inputManager, Callback<float>* callback,
                                         sf::Mouse::Button mouseButton);
     MouseButtonUnidirectionalRangeInput(const MouseButtonUnidirectionalRangeInput&) = delete;
     MouseButtonUnidirectionalRangeInput(MouseButtonUnidirectionalRangeInput&&) = delete;
